requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.4.0
Flask==3.0.3
frozenlist==1.4.1
git-filter-repo==2.45.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
mutagen==1.47.0
openai==0.28.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.0
pydantic_core==2.23.2
PyNaCl==1.5.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.6
spotipy==2.24.0
tk==0.1.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==13.0.1
Werkzeug==3.0.4
yarl==1.9.7
youtube-dl==2021.12.17
yt-dlp==2024.8.6
matplotlib==3.5.2

